Minor tweaks
1) Finish interface of custom kernel multi_innerProduct,

2) Calculate the local memory usage correctly taking into account
the SM count of the GPU,

3) Handle the case where in the bisection routine no candidate
kernel was found, i.e. no kernel reports to use any local memory.

+++ include/qdp_globalfuncs_subtype.h
@@ -140,7 +140,7 @@ sum( const OSubLattice<T>& s1 )
   while (actsize > 0) {
 
     unsigned numThreads = DeviceParams::Instance().getMaxBlockX();
-    while ((numThreads*sizeof(T2) > DeviceParams::Instance().getMaxSMem()) || (numThreads > actsize)) {
+    while ((numThreads*sizeof(T2) > DeviceParams::Instance().getMaxSMem()) || (numThreads > (unsigned)actsize)) {
       numThreads >>= 1;
     }
     unsigned numBlocks=(int)ceil(float(actsize)/numThreads);
